Change the behaviour of cleaning the old cache. Add manual command.

+++ launch.py
@@ -205,8 +205,7 @@ os.environ['GRADIO_TEMP_DIR'] = gradio_cache
 # Delete old data
 import shutil
 try:
-    # Yownas being paranoid
-    if gradio_cache.endswith('ruinedfooocus_cache'):
+    if args.clean-cache and gradio_cache.endswith('ruinedfooocus_cache'):
         shutil.rmtree(gradio_cache)
 except FileNotFoundError:
     pass
